Rename FetchAllTaxons and make it a model rather than a service
It's behaviour isn't really typical of a service object - it just wraps
responses from an API in POROs and provides a couple of convenience
methods for accessing them. This commit gives it a slightly more
descriptive name and moves it into the models folder.
